+++ src/parser/instruction/stype.rs
@@ -9,8 +9,8 @@ impl SType {
     pub fn execute(&self, cpu: &mut Processor, rs1: usize, rs2: usize, imm: i32) {
         match self {
             SType::Sb => {
-                let val1 = cpu.registers[rs1 as usize];
-                let val2 = cpu.registers[rs2 as usize];
+                let val1 = cpu.registers[rs1];
+                let val2 = cpu.registers[rs2];
                 let address = (val1 + imm) as usize;
                 cpu.memory.insert(address, val2 as u8);
             }
